requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.6.1
mangum==0.17.0
python-multipart==0.0.12
